LPD-39812 Merged commerceCatalogPage with commerceThemeMiniumPage and moved tests
